Add password strength meter and feedback messages
The feedback messages come from running the password through the
zxcvbn JS library while the password meter itself uses the zxcvbn-hs
library, via an XHR POST. This is because 1) the two libraries give
differing results, 2) we must validate at least on the backend, hence
zxcvbn-hs, and 3) the zxcvbn-hs library does not provide feedback
messages.

As was already the case, yesod apps which use this subsite must
filter request bodies, or not log them at all, in order to prevent
logging of plaintext passwords.