Replace 'scrypt' with 'password'.
The 'scrypt' package has not been maintained since 2011. This commit
removes that package in favor of 'password', which is being actively
maintained. No functionality of the program has been changed. This helps
us further our goal of running our development environment natively on
M1 Macs without the need of an emulator.

A following commit will come to override 'cryptonite' so that
aarch64-darwin will be fully supported. As of this, only x86-64_darwin
works.